Move from init to models file.
