+++ wiki/plugins/notifications/models.py
@@ -3,8 +3,7 @@ from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext_lazy as _
 from django.db.models import signals
 
-from django_notify import notify
-from django_notify.models import Subscription
+from django_notify.models import notify, Subscription
 
 from wiki import models as wiki_models
 from wiki.models.pluginbase import ArticlePlugin
